[adapters/web] implemented the distance parameter to filter [application/coordinate] implemented the distance parameter to filter and fixed the values calculations

+++ application/models/coordinate.go
@@ -3,12 +3,12 @@ package models
 type Points []Coordinate
 
 type Coordinate struct {
-	X int `json:"x"`
-	Y int `json:"y"`
+	X int64 `json:"x"`
+	Y int64 `json:"y"`
 }
 
 type Way struct {
 	From     Coordinate `json:"from"`
 	To       Coordinate `json:"to"`
-	Distante int        `json:"distance"`
+	Distante int64      `json:"distance"`
 }
